add common project
