+++ Speckle.Sdk.Testing/Framework/DummySendServerObjectManager.cs
@@ -3,7 +3,7 @@ using Speckle.Sdk.Serialisation.V2;
 using Speckle.Sdk.Serialisation.V2.Send;
 using Speckle.Sdk.Transports;
 
-namespace Speckle.Sdk.Serialization.Tests.Framework;
+namespace Speckle.Sdk.Testing.Framework;
 
 public class DummySendServerObjectManager(ConcurrentDictionary<string, string> savedObjects) : IServerObjectManager
 {
