+++ Speckle.Sdk.Testing/Global.cs
@@ -1,0 +1,1 @@
+﻿[assembly: CollectionBehavior(DisableTestParallelization = true)]
