+++ Speckle.Sdk.Testing/Speckle.Sdk.Testing.csproj
@@ -1,0 +1,19 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+      <TargetFramework>net8.0</TargetFramework>
+      <IsTestProject>true</IsTestProject>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <PackageReference Include="Microsoft.NET.Test.Sdk"  />
+      <PackageReference Include="Verify.Quibble" />
+      <PackageReference Include="Verify.Xunit" />
+      <PackageReference Include="xunit.runner.visualstudio"/>
+    </ItemGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\src\Speckle.Sdk\Speckle.Sdk.csproj" />
+    </ItemGroup>
+
+</Project>
