+++ tests/Speckle.Sdk.Serialization.Tests/Module.cs
@@ -1,0 +1,13 @@
+﻿using System.Runtime.CompilerServices;
+using Speckle.Sdk.Testing;
+
+namespace Speckle.Sdk.Serialization.Tests;
+
+public static class Module
+{
+  [ModuleInitializer]
+  public static void Initialize()
+  {
+    SpeckleVerify.Initialize();
+  }
+}
