+++ tests/Speckle.Sdk.Serialization.Tests/Speckle.Sdk.Serialization.Tests.csproj
@@ -4,21 +4,17 @@
     <TargetFramework>net8.0</TargetFramework>
     <IsTestProject>true</IsTestProject>
     <PolySharpExcludeGeneratedTypes>System.Runtime.CompilerServices.IsExternalInit;System.Runtime.CompilerServices.RequiresLocationAttribute</PolySharpExcludeGeneratedTypes>
-    <RootNamespace>Speckle.Sdk.Serialization.Tests</RootNamespace>
   </PropertyGroup>
   
   <ItemGroup>
     <PackageReference Include="altcover" />
     <PackageReference Include="FluentAssertions" />
     <PackageReference Include="Microsoft.NET.Test.Sdk"  />
-    <PackageReference Include="Verify.Quibble" />
-    <PackageReference Include="Verify.Xunit" />
-    <PackageReference Include="xunit" />
     <PackageReference Include="xunit.runner.visualstudio"/>
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\src\Speckle.Sdk\Speckle.Sdk.csproj" />
+    <ProjectReference Include="..\..\Speckle.Sdk.Testing\Speckle.Sdk.Testing.csproj" />
     <ProjectReference Include="..\..\src\Speckle.Objects\Speckle.Objects.csproj" />
   </ItemGroup>
